requirements.txt
azure-cosmosdb-table
Flask
